// src/lib/types/users.ts
// Generated by https://quicktype.io

export interface Users {
  metadata: Metadata;
  data: Datum[];
}

export interface Datum {
  _id: string;
  first_name: string;
  last_name: string;
  client_id: string;
  email: string;
  phone: string;
  status: Status;
  createdAt: string;
}

export interface Status {
  access: Access;
}

export enum Access {
  Approved = "Approved",
  Denied = "Denied",
  InReview = "In-review",
  Pending = "Pending",
}

export interface Metadata {
  total: number;
  page: number;
  limit: number;
  pages: number;
}